py-polars-bigquery/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0", "polars>=1.3.0"]
build-backend = "maturin"

[project]
name = "polars-bigquery"
version = "0.0.1"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Rust",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "google-api-core>=2.25.1",
    "google-auth>=2.40.3",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-bigquery-storage>=2.33.1",
    "polars==1.34.0",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
    "maturin>=1.13.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-benchmark>=5.1.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.10",
]

[tool.maturin]
# Allow cargo test without failing on linking to Python.
# https://github.com/PyO3/pyo3/discussions/2271#discussioncomment-2497276
cargo-extra-args = "--features extension-module"

[tool.uv]
# https://github.com/PyO3/maturin/issues/2314#issuecomment-2488542771
# Rebuild package when any rust files change
cache-keys = [{file = "pyproject.toml"}, {file = "Cargo.toml"}, {file = "**/*.rs"}]
